fix(worldgen): 修正 BlendedNoise 随机源类型与 nextDouble 精度，新增原版黄金值门禁
## 核心缺陷：BlendedNoise 的种子派生路径与原版不符

原版 RandomState.NoiseWiringHelper.wrapNew 的 BlendedNoise 分支把**随机源本身**交给
BlendedNoise（legacy_random_source=true 时 newLegacyInstance(0L)=LegacyRandomSource，
false 时 random.fromHashOf("terrain") 得到的 Xoroshiro），由它顺序构造三层 PerlinNoise。

项目此前取 terrainRng->nextLong() 当种子再 new JavaLegacyRandom(seed)，两处错误：
  (a) 丢掉了随机流本身（只用了第一个 long）；
  (b) 换了 RNG 类型（Xoroshiro → LCG）。
结果三层 PerlinNoise 的置换表与原点偏移全部错位，base_3d_noise 整体变形。
此前无任何测试能发现——DensityAstBaselineTest 的期望值正是从这个错误实现导出的。

修复：
  - PerlinNoise 旧版构造 / initLayersLegacy 参数由 JavaLegacyRandom& 改为 IRandom&；
  - BlendedNoise 构造同样改收 IRandom&，内部 PerlinNoise 由 unique_ptr 改 shared_ptr
    以便 mapAll 复用同一噪声实例；
  - NoiseBindingVisitor 按 noise_settings.useLegacyRandomSource 分支，分别传
    JavaLegacyRandom(worldSeed) 或 positionalRandom().fromHashOf("minecraft:terrain")。

## 随机源推进语义修正

  - IRandom 新增 virtual consumeCount(i32)：默认循环 nextInt()，Xoroshiro128ppRandom
    重写为循环 nextU64()（对齐 XoroshiroRandomSource.consumeCount 重写）。
    该方法被 PerlinNoise::skipOctave → consumeCount(262) 消费，推进量按 RNG 类型分派。
    注意与 skip() 语义不同：Xoroshiro 的 skip 是 2^64 跳转，不是"前进 n 步"。
  - JavaLegacyRandom::nextDouble 由 /(f64)(1ULL<<53) 改为 Java 的 float 字面量乘法
    static_cast<f32>(k) * 1.1102230246251565E-16f。Java 的 long * float 会把 k 先拓宽为
    float（只剩 24 位有效位）再乘，写成分式除法得到的是 53 位精度的不同值。
  - Xoroshiro128ppRandom::nextDouble 恢复同样的 float 中间精度，撤销 840284777 引入的
    精度回归。

## 测试

  - 新增 BlendedNoiseVanillaGoldenTest：外部真值门禁，期望值由独立复刻 MC 1.21.11
    BlendedNoise/PerlinNoise/ImprovedNoise/Xoroshiro/Legacy/RandomSupport 的 Java 参考
    实现导出，不依赖本项目任何一次运行，因此能识别"自洽但与原版不一致"。覆盖主世界
    （Xoroshiro 流）、下界、末地（Legacy 流）三个维度 × 5 seed × 8 采样点，容差 1e-12。
  - DensityAstBaselineTest 基线同步重建：原基线值是从随机源未对齐的实现导出的，把缺陷
    固化了，故全部替换并补文档说明。该文件职责仍是 AST 重构（拓扑/常量折叠）的回归门禁。
  - PerlinNoisePerfTest / DensityAstUlpTest 适配新的构造签名。

## 效果

逐方块 parity：12.25%/13.68%/11.94% → 11.97%/11.97%/10.98%（三个交叉验证区块）。

+++ src/common/util/math/random/Xoroshiro128ppRandom.cpp
@@ -112,13 +112,18 @@ u64 Xoroshiro128ppRandom::nextU64()
 f64 Xoroshiro128ppRandom::nextDouble()
 {
     // MC XoroshiroRandomSource.nextDouble():
-    //   return (double) this.nextBits(53) * 1.1102230246251565E-16D;
+    //   return this.nextBits(53) * 1.110223E-16F;
     // nextBits(53) = nextLong() >>> (64 - 53) = nextLong() >>> 11，结果落在 [0, 2^53)。
     // Java 的 >>> 是无符号右移，C++ 需要先转 u64 再右移。
-    // 原版全程是 double 精度（long 拓宽为 double 时 val < 2^53 可精确表示），
-    // 中间不得降为 float——那会丢掉 29 位有效位，与原版产生可见偏差。
+    //
+    // 【字面量是 float 后缀，不是 double】原版常量写作 `1.110223E-16F`（float），其值恰好
+    // 等于 2^-53（2 的整数次幂在 float 里可精确表示），但 Java 的 `long * float` 会先做
+    // 二元数值提升——long 拓宽为 **float**（丢 29 位有效位）再乘，结果只有 24 位精度，
+    // 最后才拓宽为 double 返回。例：k = 2^53-1 时 `k * (float)2^-53` 得 1.0，而
+    // `(double)k * 2^-53` 得 0.9999999999999999——差 1 个 double ULP，足以让依赖严格
+    // 不等号的密度函数分叉。故必须复刻 float 中间精度：先 (f32)val，再乘 f32 常量。
     const u64 val = static_cast<u64>(nextLong()) >> 11;
-    return static_cast<f64>(val) * 1.1102230246251565E-16;
+    return static_cast<f64>(static_cast<f32>(val) * 1.1102230246251565E-16f);
 }
 
 f32 Xoroshiro128ppRandom::nextFloat()
